requirements.txt
python-telegram-bot
Pillow
python-slugify

